Fikset ProductRepository save metode, da id er ændret fra long til int

+++ src/main/java/dk/kea/projektgruppe_3_full_stack_new/Service/ProductService.java
@@ -46,7 +46,7 @@ public class ProductService {
         productRepository.save(product);
     }
 
-    public void updateProduct(Long id, Product product, MultipartFile imageFile) {
+    public void updateProduct(int id, Product product, MultipartFile imageFile) {
         if (imageFile != null && !imageFile.isEmpty()) {
             try {
                 byte[] imageBytes = imageFile.getBytes();
